requirements.txt
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
langchain==0.1.10
langchain-openai==0.0.6
openai==1.12.0
chromadb==0.4.22
tiktoken==0.5.2
nltk==3.8.1